pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcribe"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "openai-whisper",
    "rich",
]

[project.scripts]
transcribe = "transcribe:main"

[tool.hatch.build.targets.wheel]
include = ["transcribe.py"]
